Extract bundle generation into a separate class
